new about dialog moved thread code to separate file drop photo2shape_utils file as it is not necessary now
