+++ aboutdialog.py
@@ -51,7 +51,7 @@ class AboutDialog(QDialog, Ui_Dialog):
         version = cfg.get("general", "version")
 
         self.lblLogo.setPixmap(QPixmap(":/icons/photo2shape.png"))
-        self.lblVersion.setText(self.tr("Version: %s") % (version())
+        self.lblVersion.setText(self.tr("Version: %s") % (version))
         doc = QTextDocument()
         doc.setHtml(self.getAboutText())
         self.textBrowser.setDocument(doc)
@@ -69,14 +69,15 @@ class AboutDialog(QDialog, Ui_Dialog):
         else:
             localeFullName = QSettings().value("locale/userLocale", "")
 
-        localeShortName = localeFullName[ 0:2 ]
-        if localeShortName in [ "ru", "uk" ]:
+        localeShortName = localeFullName[0:2]
+        if localeShortName in ["ru", "uk"]:
             QDesktopServices.openUrl(QUrl("http://hub.qgis.org/projects/photo2shape/wiki"))
         else:
             QDesktopServices.openUrl(QUrl("http://hub.qgis.org/projects/photo2shape/wiki"))
 
     def getAboutText(self):
         return self.tr("""<p>Create point shapefile from a set of geotagged photos.</p>
+                       <p>Inspired by ImagesToShape plugin from Tim Sutton.</p>
                        <p><strong>Developers</strong>: Alexander Bruy</p>
                        <p><strong>Homepage</strong>: <a href="http://hub.qgis.org/projects/photo2shape">http://hub.qgis.org/projects/photo2shape</a></p>
                        <p>Please report bugs at <a href="http://hub.qgis.org/projects/photo2shape/issues">bugtracker</a>.</p>"""
