+++ ui/photo2shapedialogbase.ui
@@ -7,7 +7,7 @@
     <x>0</x>
     <y>0</y>
     <width>403</width>
-    <height>184</height>
+    <height>172</height>
    </rect>
   </property>
   <property name="windowTitle">
@@ -22,18 +22,14 @@
     </widget>
    </item>
    <item row="0" column="1">
-    <layout class="QHBoxLayout" name="horizontalLayout">
-     <item>
-      <widget class="QLineEdit" name="inputDirEdit"/>
-     </item>
-     <item>
-      <widget class="QPushButton" name="selectInputDirButton">
-       <property name="text">
-        <string>Browse</string>
-       </property>
-      </widget>
-     </item>
-    </layout>
+    <widget class="QLineEdit" name="inputDirEdit"/>
+   </item>
+   <item row="0" column="2">
+    <widget class="QPushButton" name="btnSelectInputDir">
+     <property name="text">
+      <string>Browse</string>
+     </property>
+    </widget>
    </item>
    <item row="1" column="0">
     <widget class="QLabel" name="label_2">
@@ -43,23 +39,12 @@
     </widget>
    </item>
    <item row="1" column="1">
-    <layout class="QHBoxLayout" name="horizontalLayout_2">
-     <item>
-      <widget class="QLineEdit" name="outputFileEdit"/>
-     </item>
-     <item>
-      <widget class="QPushButton" name="selectOutputFileButton">
-       <property name="text">
-        <string>Browse</string>
-       </property>
-      </widget>
-     </item>
-    </layout>
+    <widget class="QLineEdit" name="outputFileEdit"/>
    </item>
-   <item row="2" column="0" colspan="2">
-    <widget class="QProgressBar" name="progressBar">
-     <property name="value">
-      <number>0</number>
+   <item row="1" column="2">
+    <widget class="QPushButton" name="btnSelectOutputFile">
+     <property name="text">
+      <string>Browse</string>
      </property>
     </widget>
    </item>
@@ -70,7 +55,14 @@
      </property>
     </widget>
    </item>
-   <item row="4" column="0" colspan="2">
+   <item row="2" column="0" colspan="3">
+    <widget class="QProgressBar" name="progressBar">
+     <property name="value">
+      <number>0</number>
+     </property>
+    </widget>
+   </item>
+   <item row="4" column="1" colspan="2">
     <widget class="QDialogButtonBox" name="buttonBox">
      <property name="orientation">
       <enum>Qt::Horizontal</enum>
